refactor: move Tracers to respective files
